Fix two symbiotic bugs
These two bugs are theoretically unrelated, but happened to exactly
cancel each other out so that the code behaved correctly.  Better to
fix them both though.

The first bug was that we were recording the UP event from the slot-
selection key as part of the macro.  This is probably usually harmless,
but does waste space (an extra keystroke per macro that we don't need
to store).

The second bug is that during playback, the currentState was changed
late - in particular, it was not changed before injecting the keypresses
for macro playback.  This had the effect that the first keypress was
handled with currentState == PICKING_SLOT_FOR_PLAY, which happened to
result in that keypress being swallowed.  However, because of bug #1
above, this swallowed keypress was exactly the UP event for the slot-
selection key, which we didn't need in the first place.  Fixing the
first bug without fixing this bug would be nasty, though, as it would
result in the first keystroke of every macro playback not registering.