Fix interactions with other plugins using injected keys
Fixes #2.

At a high level, the solution is to inject our events without the
INJECTED flag.  We use a separate flag so we can identify our own
injected events.

This fix is made more complicated by the fact that the Kaleidoscope core
discards release events for keys without the INJECTED flag.  See
comments in the code for more details on the changes here.